requirements.txt
blinker==1.6.2
click==8.1.6
colorama==0.4.6
Flask==2.3.2
gunicorn==19.9.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pillow==10.3.0
segno==1.6.1
selenium==4.20.0
Werkzeug==2.3.6
